# IHM_VRAI_CUSTOM_NET/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebGuitare",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Client/GetClientByNom": {
      "get": {
        "tags": [
          "Client"
        ],
        "operationId": "GetClientByNom",
        "parameters": [
          {
            "name": "P_Nom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              }
            }
          }
        }
      }
    },
    "/api/Client/GetClientById": {
      "get": {
        "tags": [
          "Client"
        ],
        "operationId": "GetClientById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              }
            }
          }
        }
      }
    },
    "/api/Client/GetClientByAuthId": {
      "get": {
        "tags": [
          "Client"
        ],
        "operationId": "GetClientByAuthId",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_CLIENT"
                }
              }
            }
          }
        }
      }
    },
    "/api/Client/AddClient": {
      "post": {
        "tags": [
          "Client"
        ],
        "operationId": "AddClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/Client/UpdateClient": {
      "put": {
        "tags": [
          "Client"
        ],
        "operationId": "UpdateClient",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_CLIENT"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Client/DeleteClient": {
      "delete": {
        "tags": [
          "Client"
        ],
        "operationId": "DeleteClient",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Commande/GetCommandesByIdClient": {
      "get": {
        "tags": [
          "Commande"
        ],
        "operationId": "GetCommandesByIdClient",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_COMMANDE"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_COMMANDE"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/C_COMMANDE"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Commande/GetCommandeById": {
      "get": {
        "tags": [
          "Commande"
        ],
        "operationId": "GetCommandeById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              }
            }
          }
        }
      }
    },
    "/api/Commande/GetCommandeByNom": {
      "get": {
        "tags": [
          "Commande"
        ],
        "operationId": "GetCommandeByNom",
        "parameters": [
          {
            "name": "P_Nom",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_COMMANDE"
                }
              }
            }
          }
        }
      }
    },
    "/api/Commande/AddCommande": {
      "post": {
        "tags": [
          "Commande"
        ],
        "operationId": "AddCommande",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_COMMANDE"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_COMMANDE"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_COMMANDE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Guitare/GetGuitareById": {
      "get": {
        "tags": [
          "Guitare"
        ],
        "operationId": "GetGuitareById",
        "parameters": [
          {
            "name": "P_Id",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/C_GUITARE"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_GUITARE"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/C_GUITARE"
                }
              }
            }
          }
        }
      }
    },
    "/api/Guitare/AddGuitare": {
      "post": {
        "tags": [
          "Guitare"
        ],
        "operationId": "AddGuitare",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/C_GUITARE"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/C_GUITARE"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/C_GUITARE"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "C_CLIENT": {
        "type": "object",
        "properties": {
          "idClient": {
            "type": "integer",
            "format": "int32"
          },
          "nomClient": {
            "type": "string",
            "nullable": true
          },
          "prenomClient": {
            "type": "string",
            "nullable": true
          },
          "adresseClient": {
            "type": "string",
            "nullable": true
          },
          "cpClient": {
            "type": "string",
            "nullable": true
          },
          "photoClient": {
            "type": "string",
            "nullable": true
          },
          "authId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "C_COMMANDE": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "numeroCom": {
            "type": "string",
            "nullable": true
          },
          "reportagePhotoCam": {
            "type": "string",
            "nullable": true
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "guitareId": {
            "type": "integer",
            "format": "int32"
          },
          "prixTotal": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "C_GUITARE": {
        "type": "object",
        "properties": {
          "idGuitare": {
            "type": "integer",
            "format": "int32"
          },
          "mancheBoisId": {
            "type": "integer",
            "format": "int32"
          },
          "toucheBoisId": {
            "type": "integer",
            "format": "int32"
          },
          "corpBoisId": {
            "type": "integer",
            "format": "int32"
          },
          "neckMicroId": {
            "type": "integer",
            "format": "int32"
          },
          "centralMicroId": {
            "type": "integer",
            "format": "int32"
          },
          "bridgeMicroId": {
            "type": "integer",
            "format": "int32"
          },
          "vibratoId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}